Added set-specific parameter for area_mean_time_series.
