+++ acme_diags/driver/default_diags/area_mean_time_series_model_vs_obs.cfg
@@ -2,7 +2,7 @@
 sets = ["area_mean_time_series"]
 case_id = "PRECT"
 variables = ["PRECT"]
-area_mean_time_series_ref_names = ["ERA-Interim", "MERRA2", "GPCP_v2.3"]
+ref_names = ["ERA-Interim", "MERRA2", "GPCP_v2.3"]
 regions = ['global', 'land', 'ocean', '90S50S','50S20S','20S20N','20N50N','50N90N']
 
 
@@ -10,5 +10,5 @@ regions = ['global', 'land', 'ocean', '90S50S','50S20S','20S20N','20N50N','50N90
 sets = ["area_mean_time_series"]
 case_id = "TREFHT"
 variables = ["TREFHT"]
-area_mean_time_series_ref_names = ["ERA-Interim", "MERRA2"]
+ref_names = ["ERA-Interim", "MERRA2"]
 regions = ['global', 'land', 'ocean', '90S50S','50S20S','20S20N','20N50N','50N90N']
