+++ acme_diags/parser/__init__.py
@@ -1,7 +1,7 @@
 from .core_parser import CoreParser
 from .zonal_mean_2d_parser import ZonalMean2dParser
 from .meridional_mean_2d_parser import MeridionalMean2dParser
-
+from .area_mean_time_series_parser import AreaMeanTimeSeriesParser
 
 SET_TO_PARSER = {
     'zonal_mean_xy': CoreParser,
@@ -10,6 +10,6 @@ SET_TO_PARSER = {
     'lat_lon': CoreParser,
     'polar': CoreParser,
     'cosp_histogram': CoreParser,
-    'area_mean_time_series': CoreParser,
+    'area_mean_time_series': AreaMeanTimeSeriesParser,
 }
